# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gooru Serve API
  version: 0.1.0
  description: Initial API contract for the gooru serve web application stack.
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post:
      summary: Create a DB-backed browser session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated user, capabilities, CSRF token, and HttpOnly session cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      summary: Revoke the current session and clear its cookie.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "200":
          description: Session revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/me:
    get:
      summary: Return the current user, capabilities, and a fresh CSRF token.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Current session details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/change-password:
    post:
      summary: Change the current user's password.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /file-selections:
    post:
      summary: Create an immutable snapshot of the files matching a library query.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSelectionCreateRequest"
      responses:
        "201":
          description: Immutable file selection snapshot created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSelectionCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /files:
    get:
      summary: List tracked files for browser grids.
      security:
        - sessionAuth: []
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: page_token
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [added, name, modified, size, kind]
            default: added
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: include_facets
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Page of tracked files.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Untrack or physically delete a selected set of files.
      description: Accepts the same explicit-ID or query-with-exclusions selector used by bulk UI actions. Strict physical deletion is accepted only when every selected file is inside configured upload targets. After explicit confirmation, delete_or_untrack deletes managed files while only untracking files outside configured upload targets.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRemovalRequest"
      responses:
        "200":
          description: Selected file locations were removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRemovalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Physical deletion was requested for a selection containing files outside configured upload targets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/search:
    post:
      summary: Search tracked files with query state in the request body.
      description: Read-only alternative to GET /files for clients that must keep free-form library query text out of browser/proxy URL histories. Uses the same search service and pagination semantics as GET /files.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchRequest"
      responses:
        "200":
          description: Page of tracked files.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /files/tags:
    post:
      summary: Add tags to selected files.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/PreferAsync"
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagMutationRequest"
      responses:
        "200":
          $ref: "#/components/responses/TagMutation"
        "202":
          $ref: "#/components/responses/AsyncOperation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    put:
      summary: Replace tags on selected files.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/PreferAsync"
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagMutationRequest"
      responses:
        "200":
          $ref: "#/components/responses/TagMutation"
        "202":
          $ref: "#/components/responses/AsyncOperation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      summary: Remove tags from selected files.
      description: If tags is empty, all tags are removed from the selected files.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/PreferAsync"
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagMutationRequest"
      responses:
        "200":
          $ref: "#/components/responses/TagMutation"
        "202":
          $ref: "#/components/responses/AsyncOperation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /ui-state:
    post:
      summary: Seal WebUI library state into an opaque protected-mode token
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserURLState'
      responses:
        '201':
          description: Protected state token
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /upload-targets:
    get:
      summary: List configured upload targets.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Configured upload targets without filesystem paths.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadTargetsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /uploads:
    post:
      summary: Upload files into a configured upload target and import them.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/PreferAsync"
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                added_at_strategy:
                  type: string
                  enum: [queue, reverse_queue, modtime]
                  description: Override the selected upload target's added-at strategy for this submission.
                queue_time_ms:
                  type: array
                  items: { type: integer, format: int64 }
                  description: Client-captured queue timestamps in Unix milliseconds, one per file.
                queue_first_time_ms:
                  type: integer
                  format: int64
                  description: Earliest client-captured queue timestamp in this submission; used with `queue_last_time_ms` to reverse queues consistently across one-file async worker requests.
                queue_last_time_ms:
                  type: integer
                  format: int64
                  description: Latest client-captured queue timestamp in this submission; used with `queue_first_time_ms` to reverse queues consistently across one-file async worker requests.
                queue_index:
                  type: array
                  items: { type: integer, minimum: 0 }
                  description: Stable queue indexes, one per file.
                queue_total:
                  type: array
                  items: { type: integer, minimum: 1 }
                  description: Queue batch size, one per file.
                source_modtime_ms:
                  type: array
                  description: Source modification timestamps in Unix milliseconds, in the same order as `files`. Invalid or missing entries are treated as unavailable.
                  items:
                    type: integer
                    format: int64
                target_id:
                  type: string
                  description: Configured upload target ID. Defaults to the first configured target.
                conflict_policy:
                  type: string
                  enum: [skip, rename, replace]
                  default: skip
                  description: How to handle uploaded filenames that already exist in the target; omitted values default to skip.
                tags:
                  type: string
                  description: Space- or comma-separated fallback initial tags applied to files that do not provide item_tags.
                item_tags:
                  type: array
                  items:
                    type: string
                  description: Per-file initial tag strings aligned with files. Each entry accepts space- or comma-separated tags; an explicit empty entry means no tags for that file and suppresses fallback tags. Omit item_tags entirely to apply fallback tags to every file.
      responses:
        "200":
          description: Synchronous upload/import result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImportResponse"
        "202":
          $ref: "#/components/responses/AsyncOperation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tags:
    get:
      summary: List known tags.
      security:
        - sessionAuth: []
      parameters:
        - name: counts
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          required: false
          description: Maximum tags to return. Defaults to 200 and is capped at 500.
          schema:
            type: integer
            default: 200
            maximum: 500
      responses:
        "200":
          description: Tags, optionally with counts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tags/namespaces:
    get:
      summary: List tag namespaces.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Known tag namespaces.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespacesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /search/suggestions:
    post:
      summary: Return search suggestions with free-form query state in the request body.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionRequest"
      responses:
        "200":
          description: Search suggestions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      summary: Return tag autocomplete suggestions.
      security:
        - sessionAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Current token prefix. Use namespace:value prefixes such as `kind:` for value suggestions.
        - name: existing
          in: query
          required: false
          schema:
            type: string
          description: Existing query expression; already-present tags are filtered from suggestions when parseable.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Matching tag suggestions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /saved-searches:
    get:
      summary: List saved searches for the current user.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Current user's saved searches.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearchesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a saved search for the current user.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchRequest"
      responses:
        "201":
          description: Saved search created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /maintenance-jobs:
    get:
      summary: List manually runnable maintenance jobs.
      description: Returns the backend-owned catalog of maintenance jobs that an administrator can start manually from the Jobs view.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Manually runnable maintenance jobs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceJobListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /operations:
    get:
      summary: List durable background operations.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: query
          required: false
          description: Specific durable operation IDs to return. May be repeated; at most 64 unique IDs are accepted. Completed results are included for this bounded status-batch form.
          schema:
            type: array
            maxItems: 64
            items:
              type: string
          style: form
          explode: true
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          required: false
          description: Zero-based number of visible operations to skip.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Visible durable operations matching the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackgroundOperationListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      summary: Clear visible terminal durable operation history.
      description: Removes completed, failed, and canceled visible operations together with their terminal child task history. Pending/running operations, hidden implementation operations, and operations with active child tasks are retained.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "200":
          description: Terminal operation history cleared.
          content:
            application/json:
              schema:
                type: object
                required: [cleared]
                properties:
                  cleared:
                    type: integer
                    format: int64
                    minimum: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /operations/events:
    get:
      summary: Stream durable background operation change signals.
      description: Sends payload-free `operations` server-sent events after committed Jobs-visible operation or task state changes, plus an initial reconciliation signal and periodic keepalive comments. Requires an authenticated administrator session.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Long-lived server-sent event stream of Jobs invalidation signals.
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
    CSRF:
      name: X-Gooru-CSRF
      in: header
      required: true
      schema:
        type: string
      description: CSRF token returned by /auth/login or /auth/me. Required for cookie-authenticated mutating requests.
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: gooru_session
  responses:
    Unauthorized:
      description: Missing or invalid session cookie.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Request is forbidden, including failed CSRF validation or insufficient privileges.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TagMutation:
      description: Aggregate tag mutation result.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TagMutationResponse"
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
    AuthMeResponse:
      type: object
      required: [user, capabilities]
      properties:
        user:
          $ref: "#/components/schemas/User"
        capabilities:
          type: object
          required: [upload, tag, delete, admin]
          properties:
            upload:
              type: boolean
            tag:
              type: boolean
            delete:
              type: boolean
            admin:
              type: boolean
        csrf_token:
          type: string
...
    FileSelectionCreateRequest:
      type: object
      properties:
        query:
          type: string
          description: Library query to snapshot. Empty is treated as *.
...